:recycle: Fix yaml reader
