+++ Sources/Kyusu/Reader/ConfigReader.swift
@@ -7,14 +7,6 @@
 
 import Foundation
 
-public typealias Path = String
-public typealias Command = [String]
-public struct YamlConfig {
-    public let sourcePaths: [Path]
-    public let ignoredPaths: [Path]
-    public let command: Command
-}
-
 public protocol ConfigReader {
     associatedtype Config
     func read(filePath: Path) throws -> Config
