+++ Sources/Kyusu/Reader/YamlReadError.swift
@@ -1,0 +1,13 @@
+//
+//  YamlReadError.swift
+//  Commander
+//
+//  Created by Yudai Hirose on 2019/06/26.
+//
+
+import Foundation
+
+public enum YamlReadError: Error {
+    case missingYamlFormat
+}
+
