+++ Sources/Kyusu/Structure/Alias.swift
@@ -1,0 +1,11 @@
+//
+//  Alias.swift
+//  Commander
+//
+//  Created by Yudai Hirose on 2019/06/26.
+//
+
+import Foundation
+
+public typealias Path = String
+public typealias Command = [String]
